Installer updates/remember me updates
- Updating release target. Now uses new gradle/cordova method for app
signing as opposed to using the old jarsigner crap (simplifies things a
lot for us)
- Renaming a bunch of the targets in the Makefile so that we no longer
use -android as a suffix (ie, `make run-android` is now just `make
run`), the idea being that this repo will not be shared between android
and ios, so might as well remove all android-specific crap (it's now
implied).
- removing the SecureCAche plugin and all refs to it (including the
autologin.js file).

overall, this preps for turtl/tracker#8 (so we can get an apk out for
testing) and does a teeny bit of prep for turtl/tracker#89 (by ripping
out the old stay logged in junk).